+++ pkg/types/shapes/stablehlo_test.go
@@ -17,4 +17,10 @@ func TestToStableHLO(t *testing.T) {
 	if got := shape.ToStableHLO(); got != "tensor<i32>" {
 		t.Errorf("ToStableHLO() = %q, want %q", got, "tensor<i32>")
 	}
+
+	shape = Make(dtypes.Float32, 1, 10).WithUniformQuantization(dtypes.Int8, dtypes.Float32, 0.1, 0)
+	want := "tensor<1x10x!quant.uniform<i8:f32, 0.1:0>>"
+	if got := shape.ToStableHLO(); got != want {
+		t.Errorf("ToStableHLO() = %q, want %q", got, want)
+	}
 }